google-adk-agents/pyproject.toml
[project]
name = "google-adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.7.0",
    "litellm>=1.74.4",
    "python-dotenv>=1.1.1",
]
